add new test data and update tests
